+++ src/model/functions/FitnessFunction.java
@@ -6,7 +6,7 @@ import java.util.BitSet;
 import model.Individual;
 
 public abstract class FitnessFunction {
-
+	
 	public abstract String getName();
 
 	public String getComplexityInfo(int graphSize, double evaporationFactor) {
@@ -19,7 +19,7 @@ public abstract class FitnessFunction {
 
 	public abstract String getThisComplexity();
 
-	public abstract int getFittness(Individual i);
+	public abstract double getFittness(Individual i);
 
 	public abstract BitSet getOptimum(int graphSize);
 }
